fix: preserve IME input across all TUI text fields
